+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 ChangeLog
 =========
 
+3.0.0 (????-??-??)
+------------------
+
+* #26: Added an event loop implementation. Also knows as the Reactor Pattern.
+* Renamed `Promise::error` to `Promise::otherwise` to be consistent with
+  ReactPHP and Guzzle. The `error` method is kept for BC but will be removed
+  in a future version.
+
+
+2.0.1 (2014-10-06)
+------------------
+
+* Fixed: `$priority` was ignored in `EventEmitter::once` method.
+* Fixed: Breaking the event chain was not possible in `EventEmitter::once`.
+
 2.0.0 (2014-06-21)
 ------------------